+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -3,6 +3,8 @@
 ## Version 1.3.0
 
 - Implemented `FunctionAppHostManager.RestartHostIfChanges(IEnumerable<KeyValuePair<string, string>> environmentVariables)` that only restarts the function app if process environment variables has changed.
+- Extended control of starting an Azure Function App through use of the `FunctionAppHostManager`:
+  - Specify which host log message to await during startup using `FunctionAppHostSettings.HostStartedEvent`.
 
 ## Version 1.2.0